allow scene objects to release resources

+++ lib/common/shape.ts
@@ -19,6 +19,8 @@ export abstract class Shape<T extends TopoDS_Shape = TopoDS_Shape> {
 
   private meshes: SceneObjectMesh[]
 
+  private disposed = false;
+
   constructor(private shape: T) {
     this.id = randomUUID()
   }
@@ -76,7 +78,10 @@ export abstract class Shape<T extends TopoDS_Shape = TopoDS_Shape> {
   }
 
   dispose() {
-    this.shape?.delete();
+    if (!this.disposed) {
+      this.disposed = true;
+      this.shape?.delete();
+    }
   }
 
   markAsMetaShape(type?: string) {
